move workflow from lib to deployment
