+++ src/cirrus/management/commands/manage.py
@@ -1,6 +1,5 @@
 import json
 import logging
-import os
 import sys
 
 from subprocess import CalledProcessError
@@ -9,13 +8,6 @@ import click
 
 from boto3 import Session
 
-from cirrus.lib.workflow import (
-    get_collections_workflows,
-    get_item,
-    get_items,
-    get_stats,
-    get_summary,
-)
 from cirrus.management.deployment import WORKFLOW_POLL_INTERVAL, Deployment
 from cirrus.management.utils.click import (
     AliasedShortMatchGroup,
@@ -327,18 +319,13 @@ def get_payloads(
 @manage.command("list-workflows")
 @pass_deployment
 def list_workflows(deployment: Deployment):
-    "List available collections and workflows"
-
-    # Need to add CIRRUS_DATA_BUCKET to parameter store
-    # cirrus_data_bucket = f"s3://{deployment.environment['CIRRUS_DATA_BUCKET']}/catalog.json"
-    cirrus_data_bucket = os.environ["CIRRUS_DATA_BUCKET"]  # Temporary
-
-    collections_workflows = get_collections_workflows(cirrus_data_bucket)
-    click.echo(json.dumps({"workflows": collections_workflows}, indent=2))
+    """List available collections/workflows combinations"""
+    workflows = deployment.get_workflows()
+    click.echo(json.dumps(workflows, indent=2))
 
 
 @manage.command("get-workflow-summary")
-@click.argument("collection")
+@click.argument("collections")
 @click.argument("workflow_name")
 @click.option(
     "--since",
@@ -357,16 +344,14 @@ def list_workflows(deployment: Deployment):
 @pass_deployment
 def get_workflow_summary(
     deployment: Deployment,
-    collection,
+    collections,
     workflow_name,
     since,
     limit,
 ):
-    "Get item (DynamoDB record) counts by state for a workflow"
-    cirrus_state_db = deployment.environment["CIRRUS_STATE_DB"]
-    summary = get_summary(
-        cirrus_state_db,
-        collection,
+    """Get item counts by state for a collections/workflow from DynamoDB"""
+    summary = deployment.get_workflow_summary(
+        collections,
         workflow_name,
         since,
         limit,
@@ -377,14 +362,13 @@ def get_workflow_summary(
 @manage.command("get-workflow-stats")
 @pass_deployment
 def get_workflow_stats(deployment: Deployment):
-    "Get workflow state transition stats"
-    cirrus_event_db_and_table = deployment.environment["CIRRUS_EVENT_DB_AND_TABLE"]
-    stats = get_stats(cirrus_event_db_and_table)
+    """Get aggregate workflow state transition stats from Timestream"""
+    stats = deployment.get_workflow_stats()
     click.echo(json.dumps(stats, indent=2))
 
 
 @manage.command("get-workflow-items")
-@click.argument("collection")
+@click.argument("collections")
 @click.argument("workflow_name")
 @click.option("--state", default=None, help="Filter by item state")
 @click.option(
@@ -412,7 +396,7 @@ def get_workflow_stats(deployment: Deployment):
 @pass_deployment
 def get_workflow_items(
     deployment: Deployment,
-    collection,
+    collections,
     workflow_name,
     state,
     since,
@@ -421,38 +405,34 @@ def get_workflow_items(
     sort_ascending,
     sort_index,
 ):
-    """List items (DynamoDB records) for a workflow"""
-    cirrus_state_db = deployment.environment["CIRRUS_STATE_DB"]
-    result = get_items(
-        cirrus_state_db,
-        collection,
+    """Get items for a collections/workflow from DynamoDB"""
+    items_page = deployment.get_workflow_items(
+        collections,
         workflow_name,
-        state=state,
-        since=since,
-        limit=limit,
-        nextkey=nextkey,
-        sort_ascending=sort_ascending,
-        sort_index=sort_index,
+        state,
+        since,
+        limit,
+        nextkey,
+        sort_ascending,
+        sort_index,
     )
-    click.echo(json.dumps(result, indent=2))
+    click.echo(json.dumps(items_page, indent=2))
 
 
 @manage.command("get-workflow-item")
-@click.argument("collection")
+@click.argument("collections")
 @click.argument("workflow_name")
 @click.argument("itemid")
 @pass_deployment
 def get_workflow_item(
     deployment: Deployment,
-    collection,
+    collections,
     workflow_name,
     itemid,
 ):
-    """Show details for a workflow item (DynamoDB record)"""
-    cirrus_state_db = deployment.environment["CIRRUS_STATE_DB"]
-    result = get_item(
-        cirrus_state_db,
-        collection,
+    """Get individual item for a collections/workflow from DynamoDB"""
+    result = deployment.get_workflow_item(
+        collections,
         workflow_name,
         itemid,
     )
